+++ module/low-light-vision.js
@@ -32,6 +32,77 @@ Token.prototype.getLightRadius = function(units) {
   return radius;
 };
 
+const Token__updateSource = Token.prototype.updateSource;
+Token.prototype.updateSource = function({defer=false, deleted=false, noUpdateFog=false}={}) {
+  if ( CONFIG.debug.sight ) {
+    SightLayer._performance = { start: performance.now(), tests: 0, rays: 0 }
+  }
+
+  // Prepare some common data
+  const origin = this.getSightOrigin();
+  const sourceId = this.sourceId;
+  const d = canvas.dimensions;
+  const maxR = canvas.lighting.globalLight ? Math.hypot(d.sceneWidth, d.sceneHeight) : null;
+
+  // Update light source
+  const isLightSource = this.emitsLight && !this.data.hidden;
+  if ( isLightSource && !deleted ) {
+    const bright = this.getLightRadius(this.data.brightLight);
+    const dim = this.getLightRadius(this.data.dimLight);
+    this.light.initialize({
+      x: origin.x,
+      y: origin.y,
+      dim: dim,
+      bright: bright,
+      angle: this.data.lightAngle,
+      rotation: this.data.rotation,
+      color: this.data.lightColor,
+      alpha: this.data.lightAlpha,
+      animation: this.data.lightAnimation
+    });
+    canvas.lighting.sources.set(sourceId, this.light);
+    if ( !defer ) {
+      this.light.drawLight();
+      this.light.drawColor();
+    }
+  }
+  else {
+    canvas.lighting.sources.delete(sourceId);
+    if ( isLightSource && !defer ) canvas.lighting.refresh();
+  }
+
+  // Update vision source
+  const isVisionSource = this._isVisionSource();
+  if ( isVisionSource && !deleted ) {
+    //-Override token vision sources to not receive low-light bonus-
+    let dim =  maxR ?? this.getLightRadius(this.data.dimSight);
+    let bright = this.getLightRadius(this.data.brightSight);
+    if (canvas.sight.hasLowLight()) {
+      dim = dim / 2;
+      bright = bright / 2;
+    }
+    //-End change-
+    if ((dim === 0) && (bright === 0)) dim = d.size * 0.6;
+    this.vision.initialize({
+      x: origin.x,
+      y: origin.y,
+      dim: dim,
+      bright: bright,
+      angle: this.data.sightAngle,
+      rotation: this.data.rotation
+    });
+    canvas.sight.sources.set(sourceId, this.vision);
+    if ( !defer ) {
+      this.vision.drawLight();
+      canvas.sight.refresh({noUpdateFog});
+    }
+  }
+  else {
+    canvas.sight.sources.delete(sourceId);
+    if ( isVisionSource && !defer ) canvas.sight.refresh();
+  }
+}
+
 const AmbientLight__get__dimRadius = Object.getOwnPropertyDescriptor(AmbientLight.prototype, "dimRadius").get;
 Object.defineProperty(AmbientLight.prototype, "dimRadius", {
   get: function() {