+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Fix Bardic Knowledge (SKA) not existing
 - Adding and removing swim and climb speeds did not update racial skill bonuses immediately
+- Changed low-light vision to no longer affect Token Dim/Bright vision [commonly darkvision] (#365)
 
 ### Changelog